Move firebase config to src/firebase
