+++ src/Login.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { signInWithEmailAndPassword } from 'firebase/auth';
-import { auth } from '../src:firebase/src:firebase:config';
+import { auth } from './firebase/config';
 
 const Login = ({ onLogin }) => {
   const [email, setEmail] = useState('');
